cleanup of functions, moved specifics to inc file with functional comments started to build in fallbacks to SCSS classes, using "use" instead of import
