--- src/js/import-test.js
@@ -1,1 +1,0 @@
-alert('Hello World!');
