+++ src/js/scripts.js
@@ -1,2 +1,1 @@
-// testing 
-import './import-test';
+// js coming soon //
