+++ src/scss/global/_project_global.scss
@@ -1,0 +1,25 @@
+@use "./typography";
+* {
+	box-sizing: border-box;
+	padding:0;
+	margin:0;
+}
+html {
+	scroll-behavior: smooth;
+}
+body {
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+.container-cst {
+	max-width: 1220px;
+	width: 100%;
+	margin: 0 auto;
+  @media only screen and (max-width: 1215px) {
+		padding-left:20px;
+		padding-right: 20px;
+  }
+}
+img {
+  max-width:100%;
+}
