+++ src/scss/global/_wp_image_alignment.scss
@@ -1,0 +1,59 @@
+img.alignright,img.alignleft,img.aligncenter, img.alignnone {
+	display: block;
+	max-width: 100%;
+	height: auto;
+}
+img.alignright {
+  margin: 1em 0;
+}
+img.alignleft {
+  margin:1em 0;
+}
+img.alignnone {
+  margin: 1em 0;
+}
+img.aligncenter {
+  display: block; 
+  margin: 1em;
+}
+a img.alignright {
+  float:right; 
+  margin:0 0 1em 1em;
+}
+a img.alignleft {
+  float:left; 
+  margin:0 1em 1em 0;
+}
+a img.aligncenter {
+  display: block; 
+  margin-left: auto; 
+  margin-right: auto;
+}
+@media only screen and (min-width: 768px) {
+	img.alignright {
+    float:right; 
+    margin:0 0 1em 1em;
+  }
+	img.alignleft {
+    float:left; 
+    margin:0 1em 1em 0;
+  }
+	img.aligncenter {
+    display: block; 
+    margin-left: auto; 
+    margin-right: auto;
+  }
+	a img.alignright {
+    float:right; 
+    margin:0 0 1em 1em;
+  }
+	a img.alignleft {
+    float:left; 
+    margin:0 1em 1em 0;
+  }
+	a img.aligncenter {
+    display: block; 
+    margin-left: auto; 
+    margin-right: auto;
+  }
+}
