+++ template-parts/header/header-content.php
@@ -1,1 +1,8 @@
-Header here
+<div class="container-cst">
+  <div class="logo-wrapp-cst">
+    <span>Logo</span>
+  </div>
+  <div class="menu-wrapp-cst">
+    <span>Menu will go here.</span>
+  </div>
+</div>
